- Move "urlSpecificTypeChecking"-(regex-related)-code from "UrlUtils" into the new "UrlTypeChecker"-class. - Move "handleUrlChecks()"-method from "UrlUtils" to "LoadAndCheckUrls". - Reposition "getOutgoingUrls()"-code inside "PageCrawler".
