+++ src/main/java/eu/openaire/doc_urls_retriever/DocUrlsRetriever.java
@@ -4,6 +4,7 @@ import eu.openaire.doc_urls_retriever.crawler.CrawlerController;
 import eu.openaire.doc_urls_retriever.util.file.FileUtils;
 import eu.openaire.doc_urls_retriever.util.http.HttpConnUtils;
 import eu.openaire.doc_urls_retriever.util.url.LoadAndCheckUrls;
+import eu.openaire.doc_urls_retriever.util.url.UrlTypeChecker;
 import eu.openaire.doc_urls_retriever.util.url.UrlUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -105,23 +106,23 @@ public class DocUrlsRetriever
 			logger.info("From which docUrls, we were able to retrieve: " + numOfStoredDocFiles + " distinct docFiles. That's about: " + numOfStoredDocFiles * (float)100 / UrlUtils.sumOfDocUrlsFound +"%."
 					+" The un-retrieved docFiles were either belonging to already-found docUrls or they had content-issues.");
 		}
-		logger.info("About: " + UrlUtils.crawlerSensitiveDomains * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.crawlerSensitiveDomains  + " urls) were from known crawler-sensitive domains.");
-		logger.info("About: " + UrlUtils.javascriptPageUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.javascriptPageUrls + " urls) were from a JavaScript-powered domain, other than the \"sciencedirect.com\", which has dynamic links.");
-		//logger.info("About: " + UrlUtils.sciencedirectUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.sciencedirectUrls + " urls) were from the \"sciencedirect.com\"-family urls, with dynamic links.");
-		logger.info("Αbout: " + UrlUtils.elsevierUnwantedUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.elsevierUnwantedUrls + " urls) were from, or reached after redirects, the unwanted domain: \"elsevier.com\", which either doesn't provide docUrls in its docPages, or it redirects to \"sciencedirect.com\", thus being avoided to be crawled.");
-		logger.info("Αbout: " + UrlUtils.doajResultPageUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.doajResultPageUrls + " urls) were \"doaj.org/toc/\" urls, which are resultPages, thus being avoided to be crawled.");
-		logger.info("Αbout: " + UrlUtils.pagesWithHtmlDocUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.pagesWithHtmlDocUrls + " urls) were docUrls, but, in HTML, thus being avoided to be crawled.");
-		logger.info("About: " + UrlUtils.pagesRequireLoginToAccessDocFiles * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.pagesRequireLoginToAccessDocFiles + " urls) were of domains which are known to require login to access docFiles, thus, they were blocked before being connected.");
-		logger.info("About: " + UrlUtils.pagesWithLargerCrawlingDepth * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.pagesWithLargerCrawlingDepth + " urls) were docPages which have their docUrl deeper inside their server, thus being currently avoided.");
-		logger.info("About: " + UrlUtils.pangaeaUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.pangaeaUrls + " urls) were \"PANGAEA.\" with invalid form and non-docUrls in their inner links.");
-		logger.info("About: " + UrlUtils.connProblematicUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.connProblematicUrls + " urls) were pages which had connectivity problems.");
-		logger.info("About: " + UrlUtils.pagesNotProvidingDocUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.pagesNotProvidingDocUrls + " urls) were pages which are known to not provide docUrls.");
-		logger.info("About: " + UrlUtils.longToRespondUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.longToRespondUrls + " urls) were urls which belong to domains which take too long to respond.");
-		logger.info("About: " + UrlUtils.urlsWithUnwantedForm * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.urlsWithUnwantedForm + " urls) were urls which are plain-domains, have unwanted url-extensions, ect...");
-		logger.info("About: " + UrlUtils.inputDuplicatesNum * (float)100 / inputCheckedUrlNum + "% (" + UrlUtils.inputDuplicatesNum + " urls) were duplicates in the input file.");
+		logger.info("About: " + UrlTypeChecker.crawlerSensitiveDomains * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.crawlerSensitiveDomains  + " urls) were from known crawler-sensitive domains.");
+		logger.info("About: " + UrlTypeChecker.javascriptPageUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.javascriptPageUrls + " urls) were from a JavaScript-powered domain, other than the \"sciencedirect.com\", which has dynamic links.");
+		//logger.info("About: " + UrlTypeChecker.sciencedirectUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.sciencedirectUrls + " urls) were from the \"sciencedirect.com\"-family urls, with dynamic links.");
+		logger.info("Αbout: " + UrlTypeChecker.elsevierUnwantedUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.elsevierUnwantedUrls + " urls) were from, or reached after redirects, the unwanted domain: \"elsevier.com\", which either doesn't provide docUrls in its docPages, or it redirects to \"sciencedirect.com\", thus being avoided to be crawled.");
+		logger.info("Αbout: " + UrlTypeChecker.doajResultPageUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.doajResultPageUrls + " urls) were \"doaj.org/toc/\" urls, which are resultPages, thus being avoided to be crawled.");
+		logger.info("Αbout: " + UrlTypeChecker.pagesWithHtmlDocUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.pagesWithHtmlDocUrls + " urls) were docUrls, but, in HTML, thus being avoided to be crawled.");
+		logger.info("About: " + UrlTypeChecker.pagesRequireLoginToAccessDocFiles * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.pagesRequireLoginToAccessDocFiles + " urls) were of domains which are known to require login to access docFiles, thus, they were blocked before being connected.");
+		logger.info("About: " + UrlTypeChecker.pagesWithLargerCrawlingDepth * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.pagesWithLargerCrawlingDepth + " urls) were docPages which have their docUrl deeper inside their server, thus being currently avoided.");
+		logger.info("About: " + UrlTypeChecker.pangaeaUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.pangaeaUrls + " urls) were \"PANGAEA.\" with invalid form and non-docUrls in their inner links.");
+		logger.info("About: " + LoadAndCheckUrls.connProblematicUrls * (float)100 / inputCheckedUrlNum + "% (" + LoadAndCheckUrls.connProblematicUrls + " urls) were pages which had connectivity problems.");
+		logger.info("About: " + UrlTypeChecker.pagesNotProvidingDocUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.pagesNotProvidingDocUrls + " urls) were pages which are known to not provide docUrls.");
+		logger.info("About: " + UrlTypeChecker.longToRespondUrls * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.longToRespondUrls + " urls) were urls which belong to domains which take too long to respond.");
+		logger.info("About: " + UrlTypeChecker.urlsWithUnwantedForm * (float)100 / inputCheckedUrlNum + "% (" + UrlTypeChecker.urlsWithUnwantedForm + " urls) were urls which are plain-domains, have unwanted url-extensions, ect...");
+		logger.info("About: " + LoadAndCheckUrls.inputDuplicatesNum * (float)100 / inputCheckedUrlNum + "% (" + LoadAndCheckUrls.inputDuplicatesNum + " urls) were duplicates in the input file.");
 		
-		long problematicUrlsNum = UrlUtils.crawlerSensitiveDomains + UrlUtils.javascriptPageUrls /*+ UrlUtils.sciencedirectUrls*/ + UrlUtils.elsevierUnwantedUrls + UrlUtils.doajResultPageUrls + UrlUtils.pagesWithHtmlDocUrls + UrlUtils.pagesRequireLoginToAccessDocFiles
-									+ UrlUtils.pagesWithLargerCrawlingDepth + UrlUtils.pangaeaUrls + UrlUtils.urlsWithUnwantedForm + UrlUtils.connProblematicUrls + UrlUtils.pagesNotProvidingDocUrls + UrlUtils.longToRespondUrls + UrlUtils.inputDuplicatesNum;
+		long problematicUrlsNum = UrlTypeChecker.crawlerSensitiveDomains + UrlTypeChecker.javascriptPageUrls /*+ UrlTypeChecker.sciencedirectUrls*/ + UrlTypeChecker.elsevierUnwantedUrls + UrlTypeChecker.doajResultPageUrls + UrlTypeChecker.pagesWithHtmlDocUrls + UrlTypeChecker.pagesRequireLoginToAccessDocFiles
+									+ UrlTypeChecker.pagesWithLargerCrawlingDepth + UrlTypeChecker.pangaeaUrls + UrlTypeChecker.urlsWithUnwantedForm + LoadAndCheckUrls.connProblematicUrls + UrlTypeChecker.pagesNotProvidingDocUrls + UrlTypeChecker.longToRespondUrls + LoadAndCheckUrls.inputDuplicatesNum;
 		logger.info("From the " + inputCheckedUrlNum + " urls checked from the input, the " + problematicUrlsNum + " of them (about " + problematicUrlsNum * (float)100 / inputCheckedUrlNum + "%) were problematic (sum of the all of the above cases).");
 		
 		logger.info("The number of domains blocked after an \"SSL Exception\" was: " + HttpConnUtils.domainsBlockedDueToSSLException);
